Refactor billing logic and update `Hours` to `decimal`
Updated `Hours` data type from `int` to `decimal` across DTOs, entities, and API schemas for precise time tracking. Refactored billing logic in `TransactionRecordService.cs`:
- Introduced `GetBilledHours` for improved hour rounding.
- Simplified board game day-pass pricing logic.
- Removed final price rounding.

Added new API endpoints:
- `GetClientUsersCount` and `GetOrdersCount` with optional filters.

Performed code cleanup:
- Removed redundant comments and unused code.
- Updated global usings to remove `global::` prefix.
- Adjusted OpenAPI schema to reflect `decimal` changes.

+++ Application/DTOs/CategoryItemDtos.cs
@@ -75,7 +75,7 @@
         public string? SetName { get; set; }
 
         // Totals
-        public int Hours { get; set; }
+        public decimal Hours { get; set; }
         public decimal TotalPrice { get; set; }
 
         // Items inside this transaction
@@ -112,7 +112,7 @@
         public int? GameSettingId { get; set; }
         public string? GameSettingName { get; set; }
 
-        public int Hours { get; set; }
+        public decimal Hours { get; set; }
         public decimal TotalPrice { get; set; }
         public DiscountDto? Discount {get;set; }
